Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/domxpath/numbers-expected.txt




FAIL "+" operator depending on the context node assert_equals: expected 3 but got 1
FAIL "-" operator depending on the context node assert_equals: expected -1 but got 1
FAIL "*" operator depending on the context node assert_equals: expected 2 but got 0
FAIL "div" operator depending on the context node assert_equals: expected 0.5 but got Infinity
FAIL "mod" operator depending on the context node assert_equals: expected 1 but got NaN
